change the riseAngle and zenithAngle from BigDecimal to float. Add a modulus function to riseAngle so that it can cycle from 0 to 1 and then back to 0. update unit tests

+++ trunk/simLifeCoreAPI/src/main/java/barsuift/simLife/environment/Sun.java
@@ -54,9 +54,9 @@ public interface Sun extends Publisher, Persistent<SunState> {
      * </ul>
      * </p>
      */
-    public BigDecimal getRiseAngle();
+    public float getRiseAngle();
 
-    public void setRiseAngle(BigDecimal riseAngle);
+    public void setRiseAngle(float riseAngle);
 
     /**
      * Zenith angle, in percent.
@@ -68,8 +68,8 @@ public interface Sun extends Publisher, Persistent<SunState> {
      * </ul>
      * </p>
      */
-    public BigDecimal getZenithAngle();
+    public float getZenithAngle();
 
-    public void setZenithAngle(BigDecimal zenithAngle);
+    public void setZenithAngle(float zenithAngle);
 
 }
